refactor: move profile utils hook to separate file
